seperate models from db_init
