+++ alembic/env.py
@@ -3,6 +3,8 @@ from logging.config import fileConfig
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 
+from database.models import Base
+
 from alembic import context
 
 # import os
@@ -21,7 +23,7 @@ if config.config_file_name is not None:
 
 # add your model's MetaData object here
 # for 'autogenerate' support
-from src.database.db_init import Base
+
 target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
