+++ alembic/versions/8f50f2459892_add_test_table.py
@@ -1,0 +1,36 @@
+"""add test table
+
+Revision ID: 8f50f2459892
+Revises: 1b69df1f13e2
+Create Date: 2023-12-25 16:06:48.424903
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '8f50f2459892'
+down_revision: Union[str, None] = '1b69df1f13e2'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('test',
+    sa.Column('name', sa.String(), nullable=False),
+    sa.Column('id', sa.Uuid(), nullable=False),
+    sa.Column('datetime_modified', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.Column('datetime_created', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('test')
+    # ### end Alembic commands ###
